Add home component and link to android file
